+++ OSKengine/DescriptorPool.cpp
@@ -30,7 +30,7 @@ void DescriptorPool::SetLayout(DescriptorLayout* layout) {
 void DescriptorPool::Create(uint32_t maxSets) {
 	VkDescriptorPoolCreateInfo poolInfo{};
 	poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-	poolInfo.poolSizeCount = descriptorPoolSizes.GetSize();
+	poolInfo.poolSizeCount = (uint32_t)descriptorPoolSizes.GetSize();
 	poolInfo.pPoolSizes = descriptorPoolSizes.GetData();
 	poolInfo.maxSets = static_cast<uint32_t>(swapchainImageCount * maxSets);
